(#3) Made leaderboard and lvl select more robust
