+++ levels/CH1/LV2.csv
@@ -4,11 +4,11 @@
 1,1,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
 1,1,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
 1,1,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
-1,1,0,0,0,0,0,0,0,0,0,D*B1&B2,0,0,0,0,0,0,0,0,0,1,1
+1,1,S,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
+1,1,1,1,0,0,0,0,0,0,0,D*B1&B2,0,0,0,0,0,0,0,0,0,1,1
 1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
 1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1
 1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,F,F,1,1
-1,1,S,0,3,0,0,0,B*B2,0,0,0,0,0,B*B1,0,0,0,F,F,0,1,1
+1,1,0,0,3,0,0,0,B*B2,0,0,0,0,0,B*B1,0,0,0,F,F,0,1,1
 1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1
 1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1
